updated AWS keys in secret file and updated .env file.
Added the attachment functionality to the Application Record models, to
use the Paperclip helper method has_attached_file and a symbol with the
desired name of the attachment. Found in
app/models/application_record.rb"

Added the attachment functionality to the Application Record models, to
use the Paperclip helper method has_attached_file and a symbol with the
desired name of the attachment. Found in
app/models/application_record.rb